data_processing/partnet_mobility/requirements.txt
numpy
pandas
tqdm
Pillow
sapien
google-generativeai
requests
